Make float32 the only interval endpoint format
--interval-dtype defaulted to float64, but every store on disk is
float32 and both the README and the validation doc passed the flag
explicitly to override that default. The default was therefore a setting
nobody used and nobody should: omitting the flag silently built a store
twice the size for no gain.

float32 is measurably sufficient. Its worst-case resolution is 4
generations, at the oldest age in the 75-draw production store
(36,744,633 generations), against the 1,000-generation analysis bin
width; near the acceptance threshold it is 0.0001 generations. float64
would grow the endpoint arrays from 13.9 GiB to 27.8 GiB on that store to
buy precision far finer than the ages themselves are resolved.

Readers take endpoint_dtype from store metadata rather than any build
flag, so float64 stores written by earlier versions still load.

INTERVAL_STORE_IMPLEMENTATION_PLAN.md prespecified float64-by-default
with float32 as an explicit, measured deviation. That prespecification is
amended rather than dropped: the measurement was made and float32
adopted outright, and the plan now records the numbers behind it.

+++ tests/test_build_snp_interval_store.py
@@ -118,7 +118,7 @@ def test_build_union_ragged_records_counts_status_and_metadata(tmp_path):
     scratch.mkdir()
     build_interval_store(
         [first, second], output, scratch_dir=scratch,
-        interval_dtype="float32", num_buckets=5,
+        num_buckets=5,
     )
     report = validate_interval_store(output, deep=True)
     assert (report.n_snps, report.n_intervals, report.n_posterior_draws) == (3, 4, 2)
